+++ Demo/src/main/java/com/paypal/android/ui/vaultcard/VaultCardViewModel.kt
@@ -227,6 +227,5 @@ class VaultCardViewModel @Inject constructor(
                 }
             }
         }
-        authState?.let { cardClient?.completeAuthChallenge(intent, it) }
     }
 }